perf(rendering): trim renderer, attribute and option temporaries
- Member ordering compares parameter passing modes as integers, so the comparison never boxes.
- Attribute names drop their suffix while being appended, and named arguments sort in one exact-size array.
- Visible types and interfaces fill bounded arrays instead of lists.
- Comma-separated options are read as spans, allocating only the entries they keep.
- Tests cover option parsing and precedence, named-argument ordering, type ordering and constraint order.

+++ src/PublicApiSharp.Analyzers/Rendering/ApiMemberOrder.cs
@@ -134,7 +134,7 @@ internal sealed class ApiMemberOrder : IComparer<ISymbol>
 
             // Two overloads may differ only in how a parameter is passed, which its type does not
             // show. Leaving them tied would let declaration order settle it.
-            result = x[i].RefKind.CompareTo(y[i].RefKind);
+            result = ((int)x[i].RefKind).CompareTo((int)y[i].RefKind);
             if (result != 0)
             {
                 return result;
